Create settings.py
I use this separate settings file so that I can change my token in one place only or if I switch between using a BETA instance of Canvas LMS or the LIVE version it makes it easier to change URL and TOKEN in one file. In later scripts you will see how I have incorporated this as an import call at the start of my python scripts.